Moving isValidateConnection to hook file
