+++ frontend/chains/ChainGraphEditor.js
@@ -26,6 +26,7 @@ import { getDefaults } from "chains/flow/TypeAutoFields";
 import { useDebounce } from "utils/hooks/useDebounce";
 import { useAxios } from "utils/hooks/useAxios";
 import { SelectedNodeContext } from "chains/editor/SelectedNodeContext";
+import { useConnectionValidator } from "chains/hooks/useConnectionValidator";
 
 // Nodes are either a single node or a group of nodes
 // ConfigNode renders class_path specific content
@@ -223,78 +224,7 @@ const ChainGraphEditor = ({ graph, chain, setChain }) => {
     [chain, reactFlowInstance, colorMode]
   );
 
-  const isValidConnection = useCallback(
-    // Connections are allowed when the source and target types match
-    // and the target has an open connector slot. Targets may optionally
-    // support multiple connections.
-
-    (connection) => {
-      // target
-      const target = reactFlowInstance.getNode(connection.target);
-      const connectors = target.data.type.connectors;
-      let connector, expectedTypes;
-      if (connection.targetHandle === "in") {
-        expectedTypes = new Set(["chain-link"]);
-      } else {
-        connector = connectors.find((c) => c.key === connection.targetHandle);
-        expectedTypes = Array.isArray(connector?.source_type)
-          ? new Set(connector.source_type)
-          : new Set([connector.source_type]);
-      }
-      const supportsMultiple = connector?.multiple || false;
-
-      // source
-      const source = reactFlowInstance.getNode(connection.source);
-      const providedType =
-        connection.sourceHandle === "out"
-          ? "chain-link"
-          : source.data.type.type;
-
-      // connection types must match
-      // HAX: adding a special case for chain-agent connections until expectedType can be
-      //      expanded to be a set of types
-      if (
-        expectedTypes.has(providedType) ||
-        (expectedTypes.has("chain") && providedType === "agent")
-      ) {
-        const instanceEdges = reactFlowInstance.getEdges();
-        const targetEdges = instanceEdges.filter(
-          (e) =>
-            e.target === target.id && e.targetHandle === connection.targetHandle
-        );
-        const sourceEdges = instanceEdges.filter(
-          (e) =>
-            e.source === source.id && e.sourceHandle === connection.sourceHandle
-        );
-        const sourceConnected = sourceEdges.length > 0;
-        const targetConnected = targetEdges.length > 0;
-
-        if (edgeUpdate.edge) {
-          // valid when updating an edge:
-          // - if connecting to the same target/source
-          // - if connecting to an unconnected target/source
-          // - if target supports multiple connections
-          const currentSource = edgeUpdate.edge.source;
-          const currentTarget = edgeUpdate.edge.target;
-          const isSameSource = connection.source === currentSource;
-          const isSameTarget = connection.target === currentTarget;
-
-          return (
-            (isSameSource || !sourceConnected) &&
-            (isSameTarget || !targetConnected || supportsMultiple)
-          );
-        } else {
-          // valid when creating a new edge
-          // - if connecting to an unconnected target/source
-          // - if target supports multiple connections
-          return !sourceConnected && (!targetConnected || supportsMultiple);
-        }
-      }
-
-      return false;
-    },
-    [reactFlowInstance]
-  );
+  const { isValidConnection } = useConnectionValidator(edgeUpdate);
 
   const onEdgeUpdateStart = useCallback(
     (event, edge) => {
